Rename vector type header file
