+++ sdk/bare/common/usr/vector_types.h
